Refactor: reorganize predictive analysis folder structure & scripts
Moving all the scripts related to predictive analysis to a dedicated folder.
Created an archive folder to store all the old R scripts - maybe be deleted at a later stage.
Renaming of ARIMAX script to final script containing all the necessary code for the predictive analysis.
Plots related to predictive analysis are also moved to the dedicated folder.